♻️🏢 Extract loader from route
