+++ apps/admin/src/routes/_layout.show.applications.$applicationId._index/route.tsx
@@ -1,13 +1,7 @@
 import { ErrorPage, getErrorTitle } from "#core/data-display/error-page";
-import { db } from "#core/db.server";
-import { assertIsDefined } from "#core/is-defined.server";
 import { PageLayout } from "#core/layout/page";
 import { getPageTitle } from "#core/page-title";
-import { assertCurrentUserHasGroups } from "#current-user/groups.server";
-import { safeParseRouteParam, zu } from "@animeaux/zod-utils";
-import { UserGroup } from "@prisma/client";
-import type { LoaderFunctionArgs } from "@remix-run/node";
-import { json } from "@remix-run/node";
+import { zu } from "@animeaux/zod-utils";
 import type { MetaFunction } from "@remix-run/react";
 import { CardComments } from "./card-comments";
 import { CardContact } from "./card-contact";
@@ -17,70 +11,11 @@ import { CardPartnership } from "./card-partnership";
 import { CardSituation } from "./card-situation";
 import { CardSituationRefusalMessage } from "./card-situation-refusal-message";
 import { CardStructure } from "./card-structure";
+import type { loader } from "./loader.server";
 
-export async function loader({ request, params }: LoaderFunctionArgs) {
-  const currentUser = await db.currentUser.get(request, {
-    select: { groups: true },
-  });
+export { loader } from "./loader.server";
 
-  assertCurrentUserHasGroups(currentUser, [
-    UserGroup.ADMIN,
-    UserGroup.SHOW_ORGANIZER,
-  ]);
-
-  const routeParams = safeParseRouteParam(RouteParamsSchema, params);
-
-  const application = await db.show.exhibitor.application.findUnique(
-    routeParams.applicationId,
-    {
-      select: {
-        billingAddress: true,
-        billingCity: true,
-        billingCountry: true,
-        billingZipCode: true,
-        comments: true,
-        contactEmail: true,
-        contactFirstname: true,
-        contactLastname: true,
-        contactPhone: true,
-        createdAt: true,
-        desiredStandSize: true,
-        discoverySource: true,
-        id: true,
-        otherPartnershipCategory: true,
-        partnershipCategory: true,
-        proposalForOnStageEntertainment: true,
-        refusalMessage: true,
-        status: true,
-        structureActivityFields: true,
-        structureActivityTargets: true,
-        structureAddress: true,
-        structureCity: true,
-        structureCountry: true,
-        structureLegalStatus: true,
-        structureLogoPath: true,
-        structureName: true,
-        structureOtherLegalStatus: true,
-        structureSiret: true,
-        structureUrl: true,
-        structureZipCode: true,
-
-        exhibitor: {
-          select: {
-            id: true,
-            profile: { select: { name: true } },
-          },
-        },
-      },
-    },
-  );
-
-  assertIsDefined(application);
-
-  return json({ application });
-}
-
-const RouteParamsSchema = zu.object({
+export const RouteParamsSchema = zu.object({
   applicationId: zu.string().uuid(),
 });
